docs: Vercel-biased docs & update Drizzle;
I removed a few things here to help standardize things. I've also updated the
Drizzle documentation to include the `POSTGRES_PORT` and setting up the role,
database, and privileges for the role on the database in the documentation. It's
helpful for folks starting from scratch on a freshly installed PostgreSQL
instance.